divide api and ui utils file
