+++ components/hooks/useConfetti.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { brewStartedSinceMs } from "../../common/utils";
+import { brewStartedSinceMs } from "../../common/utilsApi";
 import { useLatestBrew } from "./brewHooks";
 
 export const useExplode = () => {
